# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobOut": {
        "properties": {
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "error_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Category"
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "retry_available": {
            "default": false,
            "title": "Retry Available",
            "type": "boolean"
          },
          "retry_count": {
            "title": "Retry Count",
            "type": "integer"
          },
          "source_revision": {
            "default": 0,
            "title": "Source Revision",
            "type": "integer"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageOut"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "current_stage",
          "error_category",
          "course_id",
          "retry_count",
          "error_detail",
          "stages"
        ],
        "title": "JobOut",
        "type": "object"
      },
...
      "StageOut": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "error_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Category"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "input_count": {
            "title": "Input Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output_count": {
            "title": "Output Count",
            "type": "integer"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "provider_call_count": {
            "title": "Provider Call Count",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "position",
          "status",
          "attempts",
          "error_category",
          "input_count",
          "output_count",
          "provider_call_count",
          "started_at",
          "finished_at"
        ],
        "title": "StageOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Backend Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/jobs/latest": {
      "get": {
        "description": "So a client can recover \"is this course mid-processing, paused, or\nnever started\" without already holding a job id in memory -- the only\nother way to learn one is the response of the call that created it,\nwhich a page reload or navigating away loses. Returns null (200), not\n404, when processing has never been started for this course: that is\nan ordinary state, not an error.",
        "operationId": "get_latest_job_api_v1_courses__course_id__jobs_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/JobOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Latest Job Api V1 Courses  Course Id  Jobs Latest Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/process": {
      "post": {
        "description": "Create and run a processing job for a course the caller owns.",
        "operationId": "start_processing_api_v1_courses__course_id__process_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Processing",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/retry": {
      "post": {
        "description": "Resume a paused or failed job. Stages that already succeeded are not\nre-run, so a retry never duplicates completed work.\n\nRuns in the background like start_processing (9bc8cfb) -- this used to\ncall service.run() inline and block the request for the full remaining\npipeline duration, the same false-502-under-the-proxy's-timeout failure\nmode start_processing had before that fix, just never caught here.",
        "operationId": "retry_job_api_v1_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}